+++ client/public/manifest.json
@@ -14,13 +14,8 @@
     },
     {
       "src": "iconx96.png",
-      "sizes": "64x64",
-      "type": "image/x-icon"
-    },
-    {
-      "src": "iconx96.png",
       "type": "image/png",
-      "sizes": "192x192"
+      "sizes": "96x96"
     },
     {
       "src": "logo512.png",
