--- client/src/components/workoutTable/index.js
@@ -1,1 +1,0 @@
-
