+++ client/src/index.js
@@ -22,3 +22,13 @@ if ('serviceWorker' in navigator) {
       });
   });
 }
+
+const styles = {
+  root: {
+    height: '100vh',
+    margin: '0px'
+  }
+}
+
+
+export default App;
